Secure service admin views, fix height of card

+++ services/views.py
@@ -29,9 +29,13 @@ def service_detail(request, service_id):
 
     return render(request, 'services/service_detail.html', context)
 
-
+@login_required
 def add_service(request):
     """ Add a service to the site """
+    if not request.user.is_superuser:
+        messages.error(request, 'Sorry, only site owners can do that.')
+        return redirect(reverse('home'))
+        
     if request.method == 'POST':
         form = ServiceForm(request.POST, request.FILES)
         if form.is_valid():
@@ -50,9 +54,13 @@ def add_service(request):
 
     return render(request, template, context)
 
-
+@login_required
 def edit_service(request, service_id):
     """ Edit a service in the store """
+    if not request.user.is_superuser:
+        messages.error(request, 'Sorry, only site owners can do that.')
+        return redirect(reverse('home'))
+
     service = get_object_or_404(Service, pk=service_id)
     if request.method == 'POST':
         form = ServiceForm(request.POST, request.FILES, instance=service)
@@ -74,9 +82,13 @@ def edit_service(request, service_id):
 
     return render(request, template, context)
 
-
+@login_required
 def delete_service(request, service_id):
     """ Delete a service from the store """
+    if not request.user.is_superuser:
+        messages.error(request, 'Sorry, only site owners can do that.')
+        return redirect(reverse('home'))
+
     service = get_object_or_404(Service, pk=service_id)
     service.delete()
     messages.success(request, 'Service deleted!')
